+++ config/commands.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use App\Commands\HomeCommand;
 use App\Commands\McpReaderCommand;
 use Illuminate\Console\Scheduling\ScheduleFinishCommand;
 use Illuminate\Console\Scheduling\ScheduleListCommand;
@@ -25,7 +26,7 @@ return [
     |
     */
 
-    'default' => SummaryCommand::class,
+    'default' => HomeCommand::class,
 
     /*
     |--------------------------------------------------------------------------
@@ -67,6 +68,7 @@ return [
     */
 
     'hidden' => [
+        HomeCommand::class,
         McpReaderCommand::class,
         SummaryCommand::class,
         DumpCompletionCommand::class,